feat: Move Utils Method to EntityMapper Utils Class - Meeds-io/MIPs#175
